Respond to review comments - move check to CheckHighForm.
